Refactor 2025-08-21 (#7)
* Refactor 2025-08-21

* Address copilot suggestions

* Update finalizer wrapper

+++ scripts/remove/03_cfk.sh
@@ -1,10 +1,15 @@
 #!/bin/bash
 
-set -e
+set -euo pipefail
 # set -x
 
 . ./.env
 . ./scripts/functions.sh
+
+if [[ $(kubectl get namespace | grep "${NAMESPACE}" | wc -l) -lt 1 ]]; then
+    echo "Namespace ${NAMESPACE} does not exist, skipping CFK deletion"
+    exit 0
+fi
 
 helm uninstall confluent-for-kubernetes \
     -n "${NAMESPACE}"
